Update idle.sh
Made improvements to spelunking. I will need to add options to live edit speed and value values and make the time be based on that instead of a constant.
Added mark and strafe skills to the midas rotation. Mark turns on for midas, turns off for kills. Strafe reduces cooldown of midas.
Many variables are set based on my current skill levels in the game and I have to modify them manually.
I don't like that xdotool uses my cursor to click and has to have the window active in order to do active skill rotation. I'm looking for a way to do clicks without my mouse other than putting Idle Skilling into a VM.